Split the inertness test out of the SDK-gated tracing tests
test_tracing.py's module-level importorskip("opentelemetry.sdk") also gated the fresh-interpreter test that pins spans costing nothing without a provider. That guarantee is for installs without the tracing extra, so it now lives in test_tracing_inert.py and runs in SDK-less environments; verified it passes there while test_tracing.py skips.
